pyproject.toml
[tool.poetry]
name = "reflexd"
version = "0.1.0"
description = ""
authors = ["Jimmy Zelinskie <jimmy.zelinskie+git@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
watchdog = "^0.10.3"
docopt = "^0.6.2"
bencode3 = "^0.1.0"
deluge-client = "^1.9.0"
PyYAML = "^5.3.1"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
python-language-server = "^0.35.1"
rope = "^0.17.0"
pyls-black = "^0.4.6"
pyls-isort = "^0.2.0"
pyflakes = "^2.2.0"
ipdb = "^0.13.4"


[tool.black]
exclude = '/(\.git|dist)/'

[tool.pytest.ini_options]
addopts = "-x"
log_level = "debug"
minversion = "6.0"
testpaths = ["grappa"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 100
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
